+++ src/nuki/NukiBridge.js
@@ -95,6 +95,7 @@ NukiBridge.prototype._createWebHookServer = function _createWebHookServer(log, w
         var nukiId = json.nukiId + "";
         var state = json.state;
         var batteryCritical = json.batteryCritical === true || json.batteryCritical === "true";
+        var contactClosed =  json.doorsensorState !== 3;
         var mode = json.mode;
         var lock = this._getLock(nukiId);
         if (lock == null) {
@@ -109,9 +110,9 @@ NukiBridge.prototype._createWebHookServer = function _createWebHookServer(log, w
             success : true
           };
           var isLocked = lock._isLocked(state);
-          lock._setLockCache(isLocked, batteryCritical, mode);
-          log("[INFO Nuki WebHook Server] Updated lock state from webhook to isLocked = '%s' (Nuki state '%s' ) for lock '%s' (instance id '%s') with batteryCritical = '%s' and mode = '%s'.", isLocked, state, lock.id, lock.instanceId, batteryCritical, mode);
-          lock.webHookCallback(isLocked, batteryCritical, mode);
+          lock._setLockCache(isLocked, batteryCritical, contactClosed, mode);
+          log("[INFO Nuki WebHook Server] Updated lock state from webhook to isLocked = '%s' (Nuki state '%s' ) for lock '%s' (instance id '%s') with batteryCritical = '%s', contactClosed = '%s' and mode = '%s'.", isLocked, state, lock.id, lock.instanceId, batteryCritical, contactClosed, mode);
+          lock.webHookCallback(isLocked, batteryCritical, contactClosed, mode);
           response.write(JSON.stringify(responseBody));
           response.end();
         }